More flexible API using separate tasks for model loader and exporters
TODO: finish adapting existing formar

+++ src/main/java/org/colomoto/biolqm/settings/state/StateList.java
@@ -28,7 +28,7 @@ public interface StateList {
      * @param col the component number
      * @return the corresponding value
      */
-    int get(int row, int col);
+    byte get(int row, int col);
 
     /**
      * Include or hide extra components.
